pyproject.toml
[tool.poetry]
name = "isi-technology"
version = "0.1.0"
description = ""
authors = ["Konov Ivan <i_konov@ukr.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.7"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
django-redis = "^5.4.0"
django-admin-interface = "^0.26.1"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
django-extensions = "^3.2.3"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
